Create Player_class.h
This class would create a player. It would also store the player’s health, his experience and his level. This class would also define the mechanism for leveling up the player. The class will set-up a max health for the player. This would increase as the level of the player goes up. The class also defines the damage a player can cause.