Add files via upload
Monster_Class:  This class would create a monster. The monsters have been initialized as a string based array. Each monster has been made ‘smart’. This was done by giving them a difficulty level and giving them attack damage on the basis of that difficulty level. The health of the monsters would also be based on their difficulty level. The selection of monster would be done on the basis of player level and location. A low level player would not face a monster of high difficult. Another thing to note is that monster appearance in the map would be random.